+++ client/auth/tokenStore.js
@@ -1,0 +1,14 @@
+module.exports = {
+  deleteToken: function() {
+    localStorage.removeItem('authToken');
+  },
+  storeToken: function(token) {
+    localStorage.setItem('authToken', token);
+  },
+  getToken: function() {
+    return localStorage.getItem('authToken');
+  },
+  hasToken: function() {
+    return !!this.getToken();
+  }
+};
